recommadation system/requirements.txt
langchain
langchain-openai
langchain-community
python-dotenv
streamlit
mysql-connector-python
pymysql
fastapi
uvicorn
